+++ util/analysis/helpers/fastEuclid3d.m
@@ -1,17 +1,25 @@
-function distFromSoma = fastEuclid3d(somaXYZ, targetXYZ)
-	% Fast fcn for distance between 2 points
+function D = fastEuclid3d(mainXYZ, XYZ)
+    % FASTEUCLID3D
+    %
+    % Description:
+	%   Fast fcn for Euclidean distance between 2 points
+    %
+    % Syntax:
+    %   D = fastEuclid3d(mainXYZ, XYZ);
+    %
 	% INPUTS:
-	%		somaXYZ			main location (1x3)
-	%		targetXYZ		target locations (N x 3)
+	%		mainXYZ			Main location (1x3)
+	%		XYZ             Comparison locations (N x 3)
 	% OUTPUTS:
-	%		distFromSoma	distances (N x 1)
+	%		distFromSoma	Distance(s) from main location (N x 1)
 	%
 	% 18Jun2017 - SSP - created
+    % ---------------------------------------------------------------------
 
-	somaXYZ = repmat(somaXYZ, [size(targetXYZ,1) 1]);
+	mainXYZ = repmat(mainXYZ, [size(XYZ, 1) 1]);
 
-	x = bsxfun(@minus, somaXYZ(:,1), targetXYZ(:,1));
-	y = bsxfun(@minus, somaXYZ(:,2), targetXYZ(:,2));
-	z = bsxfun(@minus, somaXYZ(:,3), targetXYZ(:,3));
+	dx = bsxfun(@minus, mainXYZ(:,1), XYZ(:,1));
+	dy = bsxfun(@minus, mainXYZ(:,2), XYZ(:,2));
+	dz = bsxfun(@minus, mainXYZ(:,3), XYZ(:,3));
 
-	distFromSoma = sqrt(x.^2 + y.^2 + z.^2);
+	D = sqrt(dx.^2 + dy.^2 + dz.^2);